// src/config/model/environment-variable.enum.ts
export enum EnvironmentVariable {
    NODE_TLS_REJECT_UNAUTHORIZED = 'NODE_TLS_REJECT_UNAUTHORIZED',
    NODE_ID = 'NODE_ID',
    LOG_LEVEL = 'LOG_LEVEL',
    LOG_FORMAT = 'LOG_FORMAT',
    POSTGRES_HOST = 'POSTGRES_HOST',
    POSTGRES_PORT = 'POSTGRES_PORT',
    POSTGRES_USER = 'POSTGRES_USER',
    POSTGRES_PASSWORD = 'POSTGRES_PASSWORD',
    POSTGRES_DATABASE = 'POSTGRES_DATABASE',
    MIDDLEWARE_HOST = 'MIDDLEWARE_HOST',
    MIDDLEWARE_PORT = 'MIDDLEWARE_PORT',
    API_VERSION = 'API_VERSION',
    POSTGRES_SSLMODE = 'POSTGRES_SSLMODE',
}